+++ src/rotation.lua
@@ -224,6 +224,9 @@ end
 
 -- Init/Reset rotation status, next tranq is the first hunter on the list
 function TranqRotate:resetRotation()
+
+    TranqRotate.lastRotationReset = GetTime()
+
     for key, hunter in pairs(TranqRotate.rotationTables.rotation) do
         hunter.nextTranq = false
         TranqRotate:refreshHunterFrame(hunter)
@@ -252,8 +255,12 @@ end
 -- Return our hunter object from name or GUID
 function TranqRotate:getHunter(searchTerm)
 
+    if (searchTerm == nil) then
+        return nil
+    end
+
     for _, hunter in pairs(TranqRotate.hunterTable) do
-        if (searchTerm ~= nil and (hunter.GUID == searchTerm or hunter.name == searchTerm)) then
+        if (hunter.GUID == searchTerm or hunter.name == searchTerm) then
             return hunter
         end
     end
@@ -549,4 +556,14 @@ function TranqRotate:getTranqFailMessage(targetName, raidIconFlags)
     )
 
     return message
+end
+
+function TranqRotate:handleResetButton()
+    TranqRotate:updateRaidStatus()
+    if (TranqRotate:isPlayerAllowedToManageRotation()) then
+        TranqRotate:resetRotation()
+        TranqRotate:sendResetBroadcast()
+    else
+        TranqRotate:printPrefixedMessage(L["RESET_UNAUTHORIZED"])
+    end
 end